refactor: 배팅 금액 validation 위치 변경
- 배팅 금액 validation 위치 BettingTable로 이동
- 베팅금액이 반환시 배팅금액(기록)이 없으면 0원 반환
